Remove obsolete hooks and fix to work with installkernel-gentoo
The installkernel-gentoo script has come a long way since this project started and many of the features like building an initramfs and adding kernels to EFI/bootloader are now provided.

The remaining features are not worth maintaining especially since it would take much additional effort to conditionally disable the replaced features in installkernel-gentoo.

For this reason most functionality is removed and only generating the configure script and cleaning up /boot and modules are kept. Additional features may be added back in in the future.